Started making Swagger documentation for all routes

+++ src/basket_item/basketItem.service.ts
@@ -1,4 +1,4 @@
-import {Injectable, NotFoundException} from "@nestjs/common";
+import {Injectable, UnauthorizedException} from "@nestjs/common";
 import {Repository} from "typeorm";
 import {BasketItem} from "../entities/basketItem.entity";
 import {InjectRepository} from "@nestjs/typeorm";
@@ -13,34 +13,30 @@ export class BasketItemService {
         private itemRepo: Repository<BasketItem>
     ) {}
 
-    async findAll(paginationDto: PaginationDto) {
+    async findAll(paginationDto: PaginationDto, userId: number) {
         return await this.itemRepo.find({
             skip: paginationDto.skip,
             take: paginationDto.limit ?? DEFAULT_PAGE_SIZE,
-        });
-    }
-
-    async findOne(id: number) {
-        const result = await this.itemRepo.findOne({
             where: {
-                id
-            }
+                userId: userId,
+            },
         });
-
-        if (!result) throw new NotFoundException();
-
-        return result;
     }
 
     async create(body: CreateBasketItemDto) {
         await this.itemRepo.save(body);
     }
 
-    async delete(id: number) {
-        await this.itemRepo.delete(id);
-    }
+    async delete(id: number, userId: number) {
+        const basketItem = await this.itemRepo.findOne({
+            where: {
+                userId: userId,
+            }
+        });
 
-    async update(id: number, body: CreateBasketItemDto) {
-        await this.itemRepo.update(id, body);
+        if (!basketItem) {
+            throw new UnauthorizedException("You don't have access to this cart!");
+        }
+        await this.itemRepo.delete(id);
     }
 }
